+++ src/components/contentPoolComponents/contentPoolStyles.ts
